This merge makes the project compile, but no functionality was tested

Conflicts:
	annis-gui/src/main/java/annis/gui/SearchWindow.java
	annis-gui/src/main/java/annis/gui/controlpanel/ControlPanel.java
	annis-service/src/main/distribution/conf/spring/CommonSqlGenerator.xml
	annis-service/src/main/java/annis/dao/SpringAnnisDao.java
	annis-service/src/main/java/annis/service/internal/QueryService.java
	annis-service/src/main/java/annis/sqlgen/TableJoinsInFromClauseSqlGenerator.java
	annis-service/src/main/java/annis/sqlgen/annopool/ApAnnotateSqlGenerator.java

+++ annis-service/src/main/java/annis/administration/DefaultAdministrationDao.java
@@ -28,6 +28,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Properties;
 import java.util.logging.Level;
 import javax.sql.DataSource;
 import org.apache.commons.io.FilenameUtils;
@@ -74,6 +75,8 @@ public class DefaultAdministrationDao implements AdministrationDao
   private boolean temporaryStagingArea;
   private String schemaVersion;
   private Map<String, String> mimeTypeMapping;
+  private Map<String, String> tableInsertSelect;
+  private Map<String, String> tableInsertFrom;
   /**
    * The name of the file and the relation containing the resolver information.
    */
@@ -156,12 +159,6 @@ public class DefaultAdministrationDao implements AdministrationDao
       + " ENCODING = 'UTF8' TEMPLATE template0");
   }
   
-  protected void setupDatabase()
-  {
-    installPlPgSql();
-    createFunctionUniqueToplevelCorpusName();
-  }
-  
   protected void installPlPgSql()
   {
     log.info("installing stored procedure language plpgsql");
@@ -186,6 +183,8 @@ public class DefaultAdministrationDao implements AdministrationDao
     log.info("creating ANNIS database schema (" + dbLayout + ")");
     executeSqlFromScript(dbLayout + "/schema.sql");
     
+    // update schema version
+    jdbcTemplate.execute("DELETE FROM repository_metadata WHERE \"name\"='schema-version'");
     jdbcTemplate.execute("INSERT INTO repository_metadata "
       + "VALUES ('schema-version', '"
       + StringUtils.replace(getSchemaVersion(), "'", "''") + "');");
@@ -244,31 +243,30 @@ public class DefaultAdministrationDao implements AdministrationDao
   }
 
   @Override
-  
   public void initializeDatabase(String host, String port, String database,
     String user, String password, String defaultDatabase, String superUser,
-    String superPassword)
+    String superPassword, boolean useSSL)
   {
-    log.info("Creating Annis database and user.");
     // connect as super user to the default database to create new user and database
-    setDataSource(createDataSource(host, port,
-      defaultDatabase, superUser, superPassword));
-
-    dropDatabase(database);
-    dropUser(user);
-    createUser(user, password);
-    createDatabase(database);
-
-
-    // switch to new database, but still as super user to install stored procedure compute_rank_level
-    setDataSource(createDataSource(host, port, database,
-      superUser, superPassword));
-    setupDatabase();
-
+    if(superPassword != null)
+    {      
+      log.info("Creating Annis database and user.");
+      setDataSource(createDataSource(host, port,
+      defaultDatabase, superUser, superPassword, useSSL));
+    
+      dropDatabase(database);
+      dropUser(user);
+      createUser(user, password);
+      createDatabase(database);
+      
+      installPlPgSql();
+    }
+    
     // switch to new database as new user for the rest
     setDataSource(createDataSource(host, port, database,
-      user, password));
-
+      user, password, useSSL));
+    
+    createFunctionUniqueToplevelCorpusName();
     createSchema();
     createSchemaIndexes();
     populateSchema();
@@ -276,16 +274,24 @@ public class DefaultAdministrationDao implements AdministrationDao
   
   private DataSource createDataSource(String host, String port,
     String database,
-    String user, String password)
+    String user, String password, boolean useSSL)
   {
+    Properties props = new Properties();
     String url = "jdbc:postgresql://" + host + ":" + port + "/" + database;
-
+    
+    if(useSSL)
+    {
+      props.put("ssl", "true");
+    }
     // DriverManagerDataSource is deprecated
     // return new DriverManagerDataSource("org.postgresql.Driver", url, user, password);
 
+    props.put("user", user);
+    props.put("password", password);
+    
     // why is this better?
     // XXX: how to construct the datasource?    
-    return new SimpleDriverDataSource(new Driver(), url, user, password);
+    return new SimpleDriverDataSource(new Driver(), url, props);
   }
 
   
@@ -311,6 +317,7 @@ public class DefaultAdministrationDao implements AdministrationDao
 //    if (true) return;
 
     adjustRankPrePost();
+    adjustTextId();
     long corpusID = updateIds();
     
     importBinaryData(path);
@@ -376,16 +383,16 @@ public class DefaultAdministrationDao implements AdministrationDao
     {
       if (table.equalsIgnoreCase(FILE_RESOLVER_VIS_MAP))
       {
+        BufferedReader bReader = null;
         try
         {
 
           // count cols for detecting old resolver_vis_map table format
           File resolver_vis_tab = new File(path, table + ".tab");
           
-          BufferedReader bReader = new BufferedReader(
+          bReader = new BufferedReader(
             new InputStreamReader(new FileInputStream(resolver_vis_tab), "UTF-8"));
           String firstLine = bReader.readLine();
-          bReader.close();
           
           int cols = 9; // default number
           if (firstLine != null)
@@ -408,7 +415,7 @@ public class DefaultAdministrationDao implements AdministrationDao
               sb.append("\"element\"    varchar, ");
               sb.append("\"vis_type\"   varchar NOT NULL, ");
               sb.append("\"display_name\"   varchar NOT NULL, ");
-              sb.append("\"order\" bigint default '0', ");
+              sb.append("\"order\" integer default '0', ");
               sb.append("\"mappings\" varchar");
               sb.append(");");
               
@@ -468,6 +475,20 @@ public class DefaultAdministrationDao implements AdministrationDao
         {
           log.error("could not read {}", table + ".tab", e);
         }
+        finally
+        {
+          if(bReader != null)
+          {
+            try
+            {
+              bReader.close();
+            }
+            catch (IOException ex)
+            {
+              log.error(null, ex);
+            }
+          }
+        }
       }
       else if (table.equalsIgnoreCase("node"))
       {
@@ -483,14 +504,14 @@ public class DefaultAdministrationDao implements AdministrationDao
   
   private void bulkImportNode(String path)
   {
+    BufferedReader reader = null;
     try
     {
       // check column number by reading first line
       File nodeTabFile = new File(path, "node.tab");
-      BufferedReader reader =
+      reader =
         new BufferedReader(new InputStreamReader(new FileInputStream(nodeTabFile), "UTF-8"));
       String firstLine = reader.readLine();
-      reader.close();
       
       
       int columnNumber = firstLine == null ? 13 : 
@@ -510,8 +531,8 @@ public class DefaultAdministrationDao implements AdministrationDao
           "CREATE TEMPORARY TABLE _tmpnode"
           + "\n(\n"
           + "id bigint,\n"
-          + "text_ref bigint,\n"
-          + "corpus_ref bigint,\n"
+          + "text_ref integer,\n"
+          + "corpus_ref integer,\n"
           + "namespace varchar(100),\n"
           + "name varchar(100),\n"
           + "\"left\" integer,\n"
@@ -542,6 +563,20 @@ public class DefaultAdministrationDao implements AdministrationDao
     catch (IOException ex)
     {
       log.error(null, ex);
+    }
+    finally
+    {
+      if(reader != null)
+      {
+        try
+        {
+          reader.close();
+        }
+        catch (IOException ex)
+        {
+          log.error(null, ex);
+        }
+      }
     }
   }
   
@@ -676,8 +711,17 @@ public class DefaultAdministrationDao implements AdministrationDao
   {
     log.info("updating pre and post order in _rank");
     executeSqlFromScript("adjustrankprepost.sql");
-    log.info("analyzing rank");
+    log.info("analyzing _rank");
     jdbcTemplate.execute("ANALYZE " + tableInStagingArea("rank"));
+  }
+  
+  protected void adjustTextId()
+  {
+    log.info("updating id in _text and text_ref in _node");
+    executeSqlFromScript("adjusttextid.sql");
+    log.info("analyzing _node and _text");
+    jdbcTemplate.execute("ANALYZE " + tableInStagingArea("text"));
+    jdbcTemplate.execute("ANALYZE " + tableInStagingArea("node"));
   }
 
   /**
@@ -740,15 +784,27 @@ public class DefaultAdministrationDao implements AdministrationDao
       if (numOfEntries > 0)
       {
         StringBuilder sql = new StringBuilder();
+  
+        String predefinedFrom = 
+          tableInsertFrom == null ? null : tableInsertFrom.get(table);
+        String predefinedSelect = 
+          tableInsertSelect == null ? null : tableInsertSelect.get(table);
         
-        if (table.equalsIgnoreCase(FILE_RESOLVER_VIS_MAP))
+        if(predefinedFrom != null || predefinedSelect != null)
         {
+          if(predefinedFrom == null)
+          {
+            predefinedFrom = predefinedSelect;
+          }
+          
           sql.append("INSERT INTO ");
           sql.append(table);
-          //FIXME DIRTY!!! find a better way instead of naming the column-names in code 
-          sql.append(
-            "(corpus, version, namespace, element, vis_type, display_name, visibility, \"order\", mappings)");
-          sql.append(" (SELECT corpus, version, namespace, element, vis_type, display_name, visibility::resolver_visibility, \"order\", mappings FROM ");
+          sql.append(" ( ");
+          sql.append(predefinedSelect);
+          
+          sql.append(" ) (SELECT ");
+          sql.append(predefinedFrom);
+          sql.append(" FROM ");
           sql.append(tableInStagingArea(table)).append(")");
         }
         else
@@ -932,7 +988,7 @@ public class DefaultAdministrationDao implements AdministrationDao
   public void storeUserConfig(AnnisUserConfig config)
   {
     String sqlUpdate = "UPDATE user_config SET config=?::json WHERE id=?";
-    String sqlInsert = "INSERT INTO user_config(id, config) VALUES(?,?::json)";
+    String sqlInsert = "INSERT INTO user_config(id, config) VALUES(?,?)";
     try
     {
       String jsonVal = jsonMapper.writeValueAsString(config);
@@ -1243,4 +1299,25 @@ public class DefaultAdministrationDao implements AdministrationDao
   {
     this.mimeTypeMapping = mimeTypeMapping;
   }
+
+  public Map<String, String> getTableInsertSelect()
+  {
+    return tableInsertSelect;
+  }
+
+  public void setTableInsertSelect(Map<String, String> tableInsertSelect)
+  {
+    this.tableInsertSelect = tableInsertSelect;
+  }
+
+  public Map<String, String> getTableInsertFrom()
+  {
+    return tableInsertFrom;
+  }
+
+  public void setTableInsertFrom(Map<String, String> tableInsertFrom)
+  {
+    this.tableInsertFrom = tableInsertFrom;
+  }
+  
 }
